feat: add possibility to set rating
refers to GRP-979

+++ internal/service/driver.go
@@ -11,7 +11,7 @@ type DriverRepo interface {
 	GetDriverById(ctx context.Context, id string) (*model.Driver, error)
 	UpdateDriverById(ctx context.Context, driver model.Driver) error
 	DeleteDriverById(ctx context.Context, id string) error
-	SetRaitingById(ctx context.Context, id string, raiting float32) error
+	SetRaitingById(ctx context.Context, id string, raiting int64) error
 }
 
 type DriverService struct {
@@ -35,6 +35,6 @@ func (Driver *DriverService) DeleteProfile(ctx context.Context, id string) error
 	return Driver.DeleteDriverById(ctx, id)
 }
 
-func (Driver *DriverService) SetRaiting(ctx context.Context, id string, raiting float32) error {
+func (Driver *DriverService) SetRaiting(ctx context.Context, id string, raiting int64) error {
 	return Driver.SetRaitingById(ctx, id, raiting)
 }
