+++ internal/service/mocks/mock_user.go
@@ -65,7 +65,7 @@ func (mr *MockDriverRepoMockRecorder) GetDriverById(arg0, arg1 interface{}) *gom
 }
 
 // SetRaitingById mocks base method.
-func (m *MockDriverRepo) SetRaitingById(arg0 context.Context, arg1 string, arg2 float32) error {
+func (m *MockDriverRepo) SetRaitingById(arg0 context.Context, arg1 string, arg2 int64) error {
 	m.ctrl.T.Helper()
 	ret := m.ctrl.Call(m, "SetRaitingById", arg0, arg1, arg2)
 	ret0, _ := ret[0].(error)
